Add custom assertion and abortion macros
There are two main macros added here

DECAF_ASSERT(<cond>, <message>);
An assertion which on failure prints out the location, message and
the current program stack trace. Note that the message is optional.

DECAF_ABORT(<message>);
Aborts the program with a message, location, and the current program
stack trace. Like with assert, the message is also optional.

Finally, there are a few preset variants of abort which use a custom
message and should be used when appropriate

DECAF_UNIMPLEMENTED();
DECAF_UNREACHABLE();

Note that at some point in the future DECAF_UNREACHABLE may change to
be an optimization hint in release mode.