Implement main
This changes decaf-bin to be an executable which takes a llvm bitcode
file along with a method within that bitcode file and then runs that
method symbolically.

This means that the executable should be run like this
decaf <bitcode.bc> <test-method>

At this point I haven't included any documentation as to this within
the executable itself. I expect that we'll change the command line
interface in the future (although maybe not as part of the prototype)
once the number of applications that we are able to run increases.