Shrink dockerfile some more (#19)
* Fix dockerfile to actually remove z3 build directory

The previous change dockerfile forgot to change back to the
parent directory before removing so it ended up trying to remove
a directory that doesn't exist.

This means that half the dockerfile (approx 2.5GB) was just the z3
build directory. Not sure what the corresponds to when compressed.

* Don't install documentation
Docker best practices recommend running apt-get with the 
--no-install-recommends flag to reduce the number of unnecessary
packages that get installed.