Allow using system z3
The libz3-dev package on debian doesn't include a cmake config.
This means that we need to write our own logic to find Z3.

We do this by writing a FindZ3.cmake file and sticking it in the
module path. I've made this one default to looking for exisiting
configs (manually installed Z3 or vcpkg).

The only downside to the current FindZ3.cmake approach is that it
needs to run a small executable to extract the Z3 version. This
won't work if we ever want to cross-compile.