Move Executor::add_failure to a separate interface
This commit moves it to a FailureTracker interface which is only
responsible for handling failures. It also adds a
PrintFailureTracker which prints out the model when a failure
occurs. This has been left as the default failure tracker for
the interpreter when no other one is provided within the ctor.