Make Z3 solver use the "default" tactic
The default Z3 solver is actually two solvers in a trenchcoat. It
has one solver that uses the "default" tactic and another one that
is incremental and uses the "smt" tactic. Now, if you use the
solver in a non-incremental tactic everything is fine. However,
as soon as you do something that looks incremental, such as
- Pushing or popping solver state
- Adding more assertions after calling check-sat
without first clearing the solver state, then it'll implicitly
switch to the incremental solver.

This is a problem. The "smt" tactic is much slower than the "default"
tactic in many cases. This can make symbolic intepretation unusable
as we end up waiting minutes to hours for a symbolic computation that
looks simple.

To address this, I've changed the solver construction to explicitly
create the solver with the "default" tactic. This forgoes incremental
solving in favour of just solving faster. For the types of programs
this prototype is likely to be run on, this should be a drastic
improvement in worst-case performance. Although, it may slow down
some other versions as well.